Simplify, explain in javadoc the intent.
