+++ maven-resolver-demos/maven-resolver-demo-snippets/src/main/java/org/apache/maven/resolver/examples/resolver/Resolver.java
@@ -42,7 +42,8 @@ import org.eclipse.aether.repository.RemoteRepository;
 import org.eclipse.aether.resolution.DependencyRequest;
 import org.eclipse.aether.resolution.DependencyResolutionException;
 import org.eclipse.aether.util.graph.visitor.DependencyGraphDumper;
-import org.eclipse.aether.util.graph.visitor.PreorderNodeListGenerator;
+import org.eclipse.aether.util.graph.visitor.NodeListGenerator;
+import org.eclipse.aether.util.graph.visitor.PreorderDependencyNodeConsumerVisitor;
 import org.eclipse.aether.util.repository.AuthenticationBuilder;
 
 /**
@@ -89,10 +90,10 @@ public class Resolver {
         System.out.println("Tree:");
         System.out.println(dump);
 
-        PreorderNodeListGenerator nlg = new PreorderNodeListGenerator();
-        rootNode.accept(nlg);
+        NodeListGenerator generator = new NodeListGenerator();
+        rootNode.accept(new PreorderDependencyNodeConsumerVisitor(generator));
 
-        return new ResolverResult(rootNode, nlg.getFiles(), nlg.getClassPath());
+        return new ResolverResult(rootNode, generator.getFiles(), generator.getClassPath());
     }
 
     public void install(Artifact artifact, Artifact pom) throws InstallationException {
